+++ src/entidades/Marca.java
@@ -1,15 +1,15 @@
 package entidades;
 
 public class Marca {
-    private long id;
+    private int id;
     private String nome;
 
-    public Marca(long id, String nome) {
+    public Marca(int id, String nome) {
         this.id = id;
         this.nome = nome;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
