+++ src/entidades/Modelo.java
@@ -1,17 +1,17 @@
 package entidades;
 
 public class Modelo {
-    private long id;
+    private int id;
     private String nome;
     private Marca marca;
 
-    public Modelo(long id, String nome, Marca marca) {
+    public Modelo(int id, String nome, Marca marca) {
         this.id = id;
         this.nome = nome;
         this.marca = marca;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
